improved generic map keeper
